cleaned up scriptables classes and (re)created scriptable objects for tiles, city centers and buildings
