+++ Assets/Classes/Factories.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 38b916a2460066c4f8f5d8c952dd2294
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
