+++ Assets/Classes/ScriptableObjects/ScriptableAnimal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd1988c35be13c04dac4a334cc8c6baf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
