+++ Assets/Classes/ScriptableObjects/ScriptableBuilding.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ed1f6ac7ce48cd4c8149d17d48f76aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
