+++ Assets/Classes/ScriptableObjects/ScriptableCityCenter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bc98343c7637d674a8cdbf7434fff40c
+guid: ffc9f0b09238cb84a9e48649a9f8aa8e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
