+++ Assets/Classes/ScriptableObjects/ScriptableProduct.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a82ac3578dfb64e43b4767031a91f8ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
