+++ Assets/Classes/ScriptableObjects/ScriptableResource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6faa37a340072394e8d2f9085f852bcb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
