+++ Assets/Classes/ScriptableObjects/ScriptableTile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90f57b6608c205d4ca15c6357f530098
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
