+++ Assets/Classes/Utilities/ProductRequirement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77df1d3abd1dff0478f02ab9de367f00
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
