+++ Assets/Classes/Utilities/ResourceProbability.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6413ddb0f27d6b4abf04d9870a8deaa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
