+++ Assets/Data/Buildings/Building_Brickhut.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 15f54bdf47604834182af3c798d63881
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
