+++ Assets/Data/Buildings/Building_HerbHarvester.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4ed1f6ac7ce48cd4c8149d17d48f76aa, type: 3}
+  m_Name: Building_HerbHarvester
+  m_EditorClassIdentifier: 
+  buildingName: 6
+  basicPrefab: {fileID: 4178760509336017182, guid: 8ef392c942eee97438154a0538a8f120, type: 3}
+  product: 5
+  productionPerWorker: 2
+  isMaxWorkersFixed: 0
+  maxWorkers: 0
+  inputProducts: []
