+++ Assets/Data/Buildings/Building_HerbHarvester.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c9a99a4ea0277fe499b8fd1a13aff015
+guid: d532f56aacbe79849b2a6cf90f8f7beb
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
