+++ Assets/Data/Buildings/Building_Kitchen.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3c2676d87a651049bd9de08816f309e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
