+++ Assets/Data/Buildings/Building_Mill.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4ed1f6ac7ce48cd4c8149d17d48f76aa, type: 3}
+  m_Name: Building_Mill
+  m_EditorClassIdentifier: 
+  buildingName: 3
+  basicPrefab: {fileID: 3898998834228104886, guid: b2c3c8a4b90996f4bb62d117ae79d3f3, type: 3}
+  product: 6
+  productionPerWorker: 3
+  isMaxWorkersFixed: 1
+  maxWorkers: 5
+  inputProducts: []
