+++ Assets/Data/Buildings/Building_Mill.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: aa02201e50ca5cb41ab5587233ee420f
+guid: ace7ef82736a9f548816435a2a98881c
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
