+++ Assets/Data/Buildings/Building_RiceFarm.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65a49095838a88844a0c861f9c28e3e1
+guid: 9d3fdd0c56c255e4ea474d26250a282f
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
