+++ Assets/Data/Buildings/Building_RootHarvester.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e5ce3d352c59f644a3cd0184f110996
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
