+++ Assets/Data/Buildings/Building_WheatFarm.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4ed1f6ac7ce48cd4c8149d17d48f76aa, type: 3}
+  m_Name: Building_WheatFarm
+  m_EditorClassIdentifier: 
+  buildingName: 1
+  basicPrefab: {fileID: 4512097725937966600, guid: 77b1592d658465a4d985aa1ddaf8f87a, type: 3}
+  product: 2
+  productionPerWorker: 3
+  isMaxWorkersFixed: 0
+  maxWorkers: 0
+  inputProducts: []
