+++ Assets/Data/Buildings/Building_Woodcutter.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9b17a507eab238d459d903fd52d5b8c0
+guid: 4a69c8b056fbf3d4494ce6d47c2ca5ff
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
