+++ Assets/Data/Buildings/CityCenter_Grassland.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: CC_Grassland
-  m_EditorClassIdentifier: Assembly-CSharp::ScriptableCityCenter
+  m_Script: {fileID: 11500000, guid: ffc9f0b09238cb84a9e48649a9f8aa8e, type: 3}
+  m_Name: CityCenter_Grassland
+  m_EditorClassIdentifier: 
   cityLocation: 0
   basicPrefab: {fileID: 1034066600723307010, guid: 86dcc28ed5d6e4e4a8c24e5c09d32cec, type: 3}
   upgradedPrefab: {fileID: 0}
