+++ Assets/Data/Buildings/CityCenter_Grassland.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4cba1db2efb9d0e439a0ddefed62790f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
