+++ Assets/Data/Map/Tile_Forest.asset
@@ -9,13 +9,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: Forest
-  m_EditorClassIdentifier: Assembly-CSharp::ScriptableTile
+  m_Script: {fileID: 11500000, guid: 90f57b6608c205d4ca15c6357f530098, type: 3}
+  m_Name: Tile_Forest
+  m_EditorClassIdentifier: 
   tileType: 1
   prefab: {fileID: 4375486808293262297, guid: cf535dcb7a833444f95ae6d75ad23812, type: 3}
   resources:
   - resourceName: 2
     resourcePrefab: {fileID: 924472514708391882, guid: faba1124b6871f5408ba613941ec42a0, type: 3}
-    spawnProbability: 0.2
+    spawnProbability: 0.3
   defaultPrefab: {fileID: 7740590973213523301, guid: b4f983a4180e3cc43bcc23b263b8d931, type: 3}
