+++ Assets/Data/Map/Tile_Forest.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: bca7ae785dd4d894398b22ad6ea50c35
-folderAsset: yes
-DefaultImporter:
+guid: 988689e9706bb1449a013a08908a044c
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
