+++ Assets/Data/Map/Tile_Grassland.asset
@@ -9,13 +9,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: Grassland
-  m_EditorClassIdentifier: Assembly-CSharp::ScriptableTile
+  m_Script: {fileID: 11500000, guid: 90f57b6608c205d4ca15c6357f530098, type: 3}
+  m_Name: Tile_Grassland
+  m_EditorClassIdentifier: 
   tileType: 0
   prefab: {fileID: 3023954715332156990, guid: bad0b8f9bcaba764db734ab87122811b, type: 3}
   resources:
   - resourceName: 1
     resourcePrefab: {fileID: 7322811874412895695, guid: 8e9be178fa677b246ab1a3f8a9366250, type: 3}
-    spawnProbability: 0.5
+    spawnProbability: 0.308
   defaultPrefab: {fileID: 0}
