+++ Assets/Scenes/SampleScene.unity
@@ -564,11 +564,11 @@ MonoBehaviour:
   columns: 5
   cellSize: 1.15
   tileListScriptable:
-  - {fileID: 11400000, guid: 65a49095838a88844a0c861f9c28e3e1, type: 2}
-  - {fileID: 11400000, guid: 07eade2962b490c4993d86fdf3938c77, type: 2}
+  - {fileID: 11400000, guid: 52d7d216a077773438d063024f245d7d, type: 2}
+  - {fileID: 11400000, guid: 988689e9706bb1449a013a08908a044c, type: 2}
   cityCenterScriptables:
-  - {fileID: 11400000, guid: aa02201e50ca5cb41ab5587233ee420f, type: 2}
-  - {fileID: 11400000, guid: c9a99a4ea0277fe499b8fd1a13aff015, type: 2}
+  - {fileID: 11400000, guid: 4cba1db2efb9d0e439a0ddefed62790f, type: 2}
+  - {fileID: 11400000, guid: c135fb0588b040748b5f1e651d45eeb2, type: 2}
 --- !u!4 &352452916
 Transform:
   m_ObjectHideFlags: 0
